fix int32 overflow for operations exceeding 2 billion keys, extend random key bit-reverse to 64-bit, fix disk avail detection when db path does not exist

+++ benchmark.h
@@ -50,7 +50,7 @@ typedef enum
 typedef struct
 {
     const char *engine_name;
-    int num_operations;
+    int64_t num_operations;
     int key_size;
     int value_size;
     int num_threads;
